refactor Enum HomeSection
